complate done pro but security is not working fine due to some problem

+++ src/main/java/com/niit/shoppingcart/homecontroller/MycartController.java
@@ -119,7 +119,7 @@ public class MycartController {
 
 	}
 	@RequestMapping("/remove/{id}")
-	public ModelAndView RemoveFromCart(@PathVariable("id")Long id ){
+	public ModelAndView RemoveFromCart(@PathVariable("id")int id ){
 		
 		System.out.println("Staring of remove Method");
 		myCart.setId(id);
